fix: Update code for situation and implementation portions of Reproduce in Python
