+++ src/data/reformDefinitionCode.js
@@ -1,30 +1,24 @@
+import { optimiseHousehold } from "../api/variables";
+import { defaultYear } from "./constants";
+
 const US_REGIONS = ["us", "enhanced_us"];
 
-export function getReproducibilityCodeBlock(metadata, policy, region, type) {
-
-  /*
-  if (!policy.reform || !policy.reform.data) {
-    console.error("getReproducibilityCodeBlock: policy parameter is malformed");
-    return;
-  }
-  */
+export function getReproducibilityCodeBlock(type, metadata, policy, region, year, householdInput=null, earningVariation=null) {
 
   // Return a series of lines, concatted into an array, 
   // generated by sub-functions
 
   return [
-    ...getHeaderCode(metadata, policy, type),
-    ...getBaselineCode(policy, region, type),
+    ...getHeaderCode(type, metadata, policy),
+    ...getBaselineCode(type, policy, region),
     ...getReformCode(policy, region),
-    /*
-    ...getSituationCode(),
-    ...getImplementationCode()
-    */
+    ...getSituationCode(type, metadata, policy, year, householdInput, earningVariation),
+    ...getImplementationCode(type, region, year)
   ];
 
 }
 
-function getHeaderCode(metadata, policy, type) {
+function getHeaderCode(type, metadata, policy) {
 
   let lines = [];
 
@@ -47,12 +41,7 @@ function getHeaderCode(metadata, policy, type) {
 
 }
 
-export function getBaselineCode(policy, region, type) {
-  console.log(policy);
-  console.log(region);
-  console.log(type);
-
-  console.log(Object.keys(policy.reform.data));
+export function getBaselineCode(type, policy, region) {
 
   // Disregard baseline code for household code
   // that doesn't contain a reform
@@ -150,6 +139,92 @@ export function getReformCode(policy, region) {
 
 }
 
+export function getSituationCode(type, metadata, policy, year, householdInput, earningVariation) {
+  
+  if (type !== "household") {
+    return [];
+  }
+
+  let householdInputCopy = JSON.parse(
+    JSON.stringify(optimiseHousehold(householdInput, metadata, true)),
+  );
+
+  for (const entityPlural of Object.keys(householdInputCopy)) {
+    for (const entity of Object.keys(householdInputCopy[entityPlural])) {
+      for (const variable of Object.keys(
+        householdInputCopy[entityPlural][entity],
+      )) {
+        if (variable !== "members") {
+          if (
+            householdInputCopy[entityPlural][entity][variable][year] === null
+          ) {
+            delete householdInputCopy[entityPlural][entity][variable];
+          }
+        }
+        if (earningVariation && variable === "employment_income") {
+          delete householdInputCopy[entityPlural][entity][variable];
+        }
+      }
+    }
+  }
+
+  if (earningVariation) {
+    householdInputCopy["axes"] = [
+      [{ name: "employment_income", count: 200, min: 0, max: 200_000 }],
+    ];
+  }
+
+  let householdJson = JSON.stringify(householdInputCopy, null, 2);
+  // It's Python-safe, so we need to make true -> True and false -> False and null -> None
+  householdJson = householdJson
+    .replace(/true/g, "True")
+    .replace(/false/g, "False")
+    .replace(/null/g, "None");
+
+  let lines = [
+    "",
+    "",
+    "situation = " + householdJson,
+    "",
+    "simulation = Simulation(",
+  ];
+
+  if (Object.keys(policy.reform.data).length) {
+    lines.push("    reform=reform,");
+  }
+
+  lines = lines.concat([
+    "    situation=situation,",
+    ")",
+    "",
+    `output = simulation.calculate("household_net_income", ${year})`,
+    "print(output)",
+  ]);
+
+  return lines;
+
+}
+
+export function getImplementationCode(type, region, timePeriod) {
+  if (type !== "policy") {
+    return [];
+  }
+
+  const isCountryUS = US_REGIONS.includes(region);
+
+  return [
+    `baseline = Microsimulation(${
+      isCountryUS ? "reform=baseline_reform" : ""
+    })`,
+    "reformed = Microsimulation(reform=reform)",
+    `baseline_person = baseline.calc("household_net_income",`,
+    `    period=${timePeriod || defaultYear}, map_to="person")`,
+    `reformed_person = reformed.calc("household_net_income",`,
+    `    period=${timePeriod || defaultYear}, map_to="person")`,
+    "difference_person = reformed_person - baseline_person",
+  ];
+}
+
 export function getStartEndDates(policy) {
 
   let earliestStart = null;
@@ -171,48 +246,4 @@ export function getStartEndDates(policy) {
     earliestStart: earliestStart,
     latestEnd: latestEnd
   });
-}
-
-export function getReformDefinitionCode(metadata, policy) {
-  let lines = [
-    "from policyengine_core.reforms import Reform",
-    "from policyengine_core.periods import instant",
-    "import pandas as pd",
-    "",
-    "",
-    "def modify_parameters(parameters):",
-  ];
-
-  if (Object.keys(policy.reform.data).length === 0) {
-    lines.pop();
-    return lines;
-  }
-
-  for (const [parameterName, parameter] of Object.entries(policy.reform.data)) {
-    for (let [instant, value] of Object.entries(parameter)) {
-      const [start, end] = instant.split(".");
-      if (value === false) {
-        value = "False";
-      } else if (value === true) {
-        value = "True";
-      }
-      lines.push(
-        `    parameters.${parameterName}.update(`,
-        `        start=instant("${start}"), stop=instant("${end}"),`,
-        `        value=${value})`,
-      );
-    }
-  }
-  lines.push("    return parameters");
-
-  lines = lines.concat([
-    "",
-    "",
-    "class reform(Reform):",
-    "    def apply(self):",
-    "        self.modify_parameters(modify_parameters)",
-    "",
-    "",
-  ]);
-  return lines;
 }
